% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{listW}
\alias{listW}
\title{List of 222 spatial weights (style = "W", zero.policy=TRUE) used in Cerqueti, R., Maranzano, P. & Mattera, R. "Spatially-clustered spatial autoregressive models with application to agricultural market concentration in Europe". arXiv preprints (<https://doi.org/10.48550/arXiv.2407.15874>)}
\format{
\describe{
\code{listW} is a list of 222 spatial weights (style = "W", zero.policy=TRUE) for the European NUTS-2 regions
}
}
\source{
{
Eurostat --  GISCO Territorial units for statistics (NUTS) (https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units/territorial-units-statistics)
}
}
\usage{
data(Data_RC_PM_RM_JABES2024)
}
\description{
The 'Data_RC_PM_RM_JABES2024' dataset was created merging information from the Eurostat regional database (<https://ec.europa.eu/eurostat/web/regions/database>).
it is a spatial dataset to replicate the results for 2020 from Cerqueti, R., Maranzano, P. & Mattera, R. "Spatially-clustered spatial autoregressive models with application to agricultural market concentration in Europe". arXiv preprints (<https://doi.org/10.48550/arXiv.2407.15874>).
Data contained in this file refer to the agricultural sector industry for 222 European regions (NUTS-2 classification) for 2020.
For more information see the database 'Economic accounts for agriculture by NUTS 2 region' (agr_r_accts, DOI:10.2908/agr_r_accts).
The file includes 6 mixed-type objects:
}
\note{
{
All source data files prepared by Paolo Maranzano (Department of Economics, Management and Statistics, University of Milano-Bicocca, Italy).
}
}
\keyword{datasets}
