% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cdisc_data.R
\name{ls_synthetic_cdisc_data}
\alias{ls_synthetic_cdisc_data}
\title{List Available Data}
\usage{
ls_synthetic_cdisc_data()
}
\description{
List Available Data
}
\examples{

ls_synthetic_cdisc_data()

}
