% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cdisc_data.R
\name{synthetic_cdisc_data}
\alias{synthetic_cdisc_data}
\title{Get Synthetic CDISC Data}
\usage{
synthetic_cdisc_data(name)
}
\arguments{
\item{name}{name of data collection to return. If \code{name = "latest"} then the newest datasets get returned.}
}
\value{
A named \code{list} containing synthetic datasets.
}
\description{
Get Synthetic CDISC Data
}
\examples{
\dontrun{
library(scda.2022)

dfs <- synthetic_cdisc_data("rcd_2022_06_27")
names(dfs)

latest_dfs <- synthetic_cdisc_data("latest")
names(latest_dfs)
}

}
