% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cdisc_data.R
\name{synthetic_cdisc_dataset}
\alias{synthetic_cdisc_dataset}
\title{Get Synthetic CDISC Dataset}
\usage{
synthetic_cdisc_dataset(archive_name, dataset_name)
}
\arguments{
\item{archive_name}{name of data collection. If \code{archive_name = "latest"} then the newest dataset gets returned.}

\item{dataset_name}{the lowercase name of the requested dataset (e.g. \code{"adsl"}).}
}
\value{
A \code{data.frame} containing synthetic data.
}
\description{
Get Synthetic CDISC Dataset
}
\examples{
\dontrun{
library(scda.2022)

adsl <- synthetic_cdisc_dataset("latest", "adsl")
}

}
