% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_cdisc_data.R
\name{synthetic_cdisc_data}
\alias{synthetic_cdisc_data}
\title{Get Synthetic CDISC Data}
\usage{
synthetic_cdisc_data(name)
}
\arguments{
\item{name}{name of data collection. If name == "latest" then the newest datasets get returned}
}
\value{
A named list with synthetic data
}
\description{
Get Synthetic CDISC Data
}
\examples{
\dontrun{
library(scda.2020)
library(scda.2021)

dfs <- synthetic_cdisc_data("rcd_2021_03_22")
names(dfs)

latest_dfs <- synthetic_cdisc_data("latest")
names(latest_dfs)
}
}
